Update screenshot-visual-diff.md
I'm certain this isn't the best, but I wanted to help make this page a little more helpful for people who would like to test the visual regression feature in Stencil. Let me know if you have any edits - this is certainly more of a brain dump than a thought out post.