Update overview.md
The docs make an assumption that the user has read the config first and knows that buildEs5 must be set to true in order for the second file under differential bundling to be generated.

Perhaps add a note or link to the config page about buildEs5. As a first time user I spent an extra half hour trying to debug why this file was not there.