docs(testing): add explicit mention of testing lib versions (#804)
state the versions of jest and puppeteer that are supported by stencil. this is the 'required' work for supporting Jest 27+. however, i took an initial stab at some rework of the testing overview page, but by no means is it "good enough" for us to mark it off our sheet of pages we want to improve. the test configuration & IDE integration could certainly use a little more love, but I think it's in a slightly better state than where it was/is today