docs(getting-started): explain cli output
explain the output of the `npx init stencil` command, to try to bridge
the gap between running the command and getting started/running with
stencil for the first time.

there's some duplication of information here. ie if we were to change
the cli output, we'd need to change the instructions here as well, which
as a landing point of sorts of our docs, i'm looking for
feedback/thoughts here