docs(browser): move browser support docs to reference section
this commit moves the pre-existing browser support docs to the
'references' section of the stencil doc site. previously, this page was
listed under the 'introduction' section, cluttering it up.

we do not add a redirect here, as the path to this page stays the same,
only it's location in the sidebar (and the source code) changes