docs(assets): rewrite assets documentation
this commit includes a full rewrite of the existing assets
documentation. it renames the documentation from 'local assets' to
'assets', and sets up a permanent redirect for external links.

this commit introduces several nuances to using output targets other
than `dist` and `www`. because this is documentation, it does not fix
the rough patches in the core product itself. rather it seeks to
acknowledge them in order to raise awareness and gain signal around
places the core compiler can be improved