docs(node): update support policy for node in v3
this commit updates the node runtime support table to state that node
v12 will not be supported in stencil v3. we do not update the table to
explicitly state that node v18 will be supported in v3 (although it's
expected to), as we haven't tested against node 18 at this time

this commit also **stubs** out the v3 release in the maintenance
schedule table. we will update this when the time comes closer to having
a firmer date