chore(docs): reorder guides table of contents (#872)
this commit reorganizes the table of contents for the guides section of
the doc site. unlike other sections, the documents in this section
aren't directly related to one another. as a result, there's no natural
'flow' between the sections. as a result, finding the guide you're
looking for requires a little extra effort to scan through the entire
contents until you find what you're looking for (today). by
alphabetizing this list, the hope is that it makes finding a specific
guide a little bit easier