fix(redirects): handles trailing slashes in redirects (#875)
add support for trailing slashes in URLs that don't use globs for
trailing slashes

remove the redirect for `docs/design-systems/` that is no longer valid
following https://github.com/ionic-team/stencil-site/pull/858

remove the redirect to `docs/build-conditionals`, which is not surfaced
in the doc site today. redirect all traffic to `docs/build-conditionals/` to
the intro page. STENCIL-438 created to investigate further