docs(resources): remove stencil announcement presentation (#880)
as a part of an initiative to clearly define the projects that the
stencil team is maintaining, the team is going through and archiving
projects that will no longer be maintained by ionic nor by the
community. one such example is https://github.com/ionic-team/stencil-present

before we archive that project, i'd like to remove it from the
resources/ page on the stencil site.

in removing the stencil component that wrapped the iframe, any css
classes that were removed were searched for in the codebase. any classes
that were only used for the removed component were also removed