docs(docs-readme): update docs-readme description of alternative CLI flag
The documentation is currently out of date, describing the use of a
`--docs-readme` CLI flag which Stencil does not support. However, the
`--docs` CLI flag does what the docs say that the `--docs-readme` flag
should do, so we can point the user towards that instead.

This was reported in ionic-team/stencil#3405

STENCIL-472