add the npm2yarn plugin
This adds a plugin which modifies a codeblock with `npm` commands in it
to have a user-selectable `npm` or `yarn` option and automatically
translates the command to `yarn`, if selected.

This commit also updates a few files to use the plugin.

See https://docusaurus.io/docs/markdown-features/code-blocks#npm2yarn-remark-plugin