docs(types): components.d.ts should be checked in (#907)
this commit adds a notice that `components.d.ts` should be checked in
to source control alongside the rest of a project's code. at the time of
this writing, stencil uses this file to complete type checking. omitting this
file can cause build steps (either locally or in continuous integration) to
fail to catch certain type errors