docs(reactive-data): update @watch documentation
this commit updates the documentation for `@Watch`. it primarily focuses
on discussing how to get the decorator to fire for class members that
are objects/arrays in order to provide clarification for how stencil
compares old/new values (by reference).

minor copy edits throughout the page have been made to improve clarity