docs(v2): create v2 versioned docs (#984)
this commit creates a v2.Y.Z set of versioned documents, by running the
`npm run docusaurus docs:version v2` command. doing so generated almost
all files found in this commit, where the contents of ./docs were copied
to ./versioned_docs/version-v2 and ./versioned_sidebars/version-v2 (and
metadata files for versions updated respectively.

the version 'v2' was specifically chosen over 'v2.22' or 'v2.0' to
demonstrate the documentation is meant to be for all of stencil v2. this
is similar to how the previous documentation site worked (one site per
version). we do not have the historical data to easily backfill versions
of stencil v2.x (and have opted not to try).

the docusaurus config was also updated in this commit such that the
contents of ./docs are no longer labeled as 'v2' (since we just created
the v2 docs.