chore(v3): move v3.0 docs to stable
this commit is the result of running `npm run docusaurus docs:version
v3.0` at the root of the repo. it generates a new set of versioned docs
for all of stencil v3.0.Z (all patch versions). as a result of this
commit, the v3.0 docs will be the default that users land on when
accessing the doc site