docs(v3): remove dist-custom-elements-bundle pages
this commit removes the `dist-custom-elements-bundle`-related
documentation from the v3.0 and next versions of the site. it does not
remove mentions of them in either version in the case of guides to
upgrade to stencil v3.0.

the path for reaching the v2-related documentation had to be updated
manually, as docusaurus complained about broken links on build once the
`custom-elements-bundle.md` file was deleted.