docs(state-tunnel): remove state-tunnel pages
this commit removes the state-tunnel related documentation that was
originally removed in https://github.com/ionic-team/stencil-site/pull/937
and accidentally lost in
https://github.com/ionic-team/stencil-site/pull/902.

the redirects created in #937 are still present. prior to applying to
this commit, they worked as intended. however, the state-tunnel docs
were reachable via the sidebar. after applying this commit, the
redirects continue to work, and the docs (being removed) are no longer
reachable