docs(angular-integration): move single component imports to FAQ
This commit moves a small section of the docs for importing individual components to the FAQ section where leveraging the `dist-custom-elements` output is discussed. This change was made because in this case users do not need to implement the Stencil loader for registering their custom elements. Instead, when using `includeImportCustomElements`, the elements are automatically registered as Angular components are bootstrapped.