chore(github): add dependabot to stencil docs
this commit adds a dependabot configuration to the stencil site to check
for updated packages used by the site itself.

at the time of this commit, the ionic-preset-classic project was
intentionally excluded from this configuration, as we expect it to be a
separate npm package in the near future.

this configuration checks weekly (as oppposed to daily in other stencil
projects) in order to see how a weekly-based notification might help
balance providing the team more time to work on stencil, rather than
dependabot updates.