fix(config): master -> main for editUrl (#1041)
this commit updates the github url for editing a page.
when we updated the primary branch in github to `main`,
we never updated this url to point to `main` as well. this didn't
appear to break anything immediately, as github would
redirect on our behalf. however, it is best we update this now,
prior to those redirects expiring