docs(support): add node 18 to support policy
this commit designates node 18 suppport in stencil 3. it also adds a
footnote that stencil 2 may be able to support node 18, as no known
issues have been reported when using the two.

the v2 version of the docs are slightly different from v3.0 and 'next',
as it omits the v3 row