docs(all): use remark-validate-links to check internal links, convert to file-path links (#1022)
This adds `remark` and `remark-validate-links` and sets us up to use them to validate that internal links (i.e. links between Markdown files) are valid. While Docusaurus already validates that such links point to URLs which are present in the rendered site, it doesn't check that headers reference in other files are present.

In order to check headers we also needed to convert all the internal links from relative links based on the final rendered URL of the destination to file-path based links which are relative paths from the source file to the destination file.

In practice, this might mean changing a link like

```md
[dist output target](/docs/distribution)
```

to something like

```md
[dist output target](../output-targets/dist.md)
```

As mentioned above, this allows `remark-validate-links` to check that any headers referenced in such a link are present in the destination file. It also has a secondary benefit that links work on Github, since they are based simply on relative location of files rather than on the URL of pages in the built site.