chore(infra): add volta to the project (#1051)
this commit adds volta to the project, allowing for the team to be using
a consistent version of node/npm when working on the site. when volta is
on the developers PATH, the version of npm/node will be set
automatically for them. this is being done to align with the rest of the
stencil developer environments found in stencil, stencil-sass, etc