chore(): regenerate package-lock.json (#1057)
When #1027 was merged the `package-lock.json` for `ionic-preset/` was changed but those changes didn't make their way to the root `package-lock.json`, so doing an `npm ci` would result in an error that the `package-lock.json` and `package.json` were out of sync, preventing installing dependencies. This fixes the issue by running `npm i`.