chore(docs): remove unused frontmatter
this commit removes unused 'contributor' frontmatter from the site.
prior to switching to docusaurus, it was used as a part of the stencil
site build to generate 'contributor bubbles' at the bottom of the page.
we no longer do this, and if we wanted to could derive this data from
git history (at the cost of calculating it).

remove it as it's unclear for new contributors if they should update
these or not.

remove it as it also conflicts with new cspell integration work, which
flags many github user names as mis-spellings