fix(ci): fix deploys, remove mv command
this commit removes the 'mv www' from the build.ci command. sometime
between 2023.03.15 and 2023.03.16, vercel deploys began to fail due to
the `build/` directory being missing after build. the build passes
locally, which gives cause to something changing on the vercel side as
far as this check goes. this `mv` is believed to be a remnant of the old
stencil site + transition to docusaurus, and no longer needed