fix(redirects): fix versioned redirects (#1071)
This PR adds functionality to the vercel.config file to include versioned docs for the current redirects. In addition, it removes pattern matching for trailing slashes and adds a config option to remove them by default.

Here is the current [pattern matching library](https://github.com/pillarjs/path-to-regexp/tree/v6.1.0) that vercel uses for its config file. Here is the [related section](https://vercel.com/docs/errors#invalid-route-source-pattern) in the vercel docs.

Unfortunately this is a blanket solution that will affect all versions (including current) of certain files, some which may not want redirects or require different destinations based on versions. I'm open to suggestions on how to mitigate this. ATM the best solution I can think of is to just handle those on a case by case basis.