docs(3.4): add v3.4 documentation
this round of documentation was created differently than other versions.
rather than using docusaurus `version`, it was manually generated by
running the following on the `HEAD` of `main`:
- `cp -R versioned_docs/version-v3.3 versioned_docs/version-v3.4`
- `cp versioned_sidebars/version-v3.3-sidebars.json versioned_sidebars/version-v3.4-sidebars.json`
- adding "v3.4" to the head of the array in `versions.json`

this was done manually as the team is currently using the `docs/`
directory for Stencil v4 staging. had we run the docusaurus versioning
command, this would have pulled v4 related changes into the v3.4 docs.

note anything merged into the v3.3 docs will need to be cherry-picked
onto this branch until it is merged. at the same time, any changes in
this branch will need to be propagated to `main`