chore(spelling): add spellcheck for the root-level .md files
This adds spell-checking for the root-level .md files (which is just
`readme.md` and `CONTRIBUTING.md`). Adding this didn't find any spelling
errors! 🎉 However, it did require adding a new ignore regex to the
cspell config to ignore GitHub emoji shortcodes (which are
colon-delimited) and adding 'Ionicons' to the wordlist.