chore(deps): update ionic-docs/preset-classic, docusaurus to v3 (#1290)
* chore(docs): mdx 3 syntax updates

* styles: site layer -> local layer

* chore(deps): bump docusaurus related deps and remove unnecessary

* Revert "chore(docs): mdx 3 syntax updates"

This reverts commit 46c16c333e10fa466aa82ae4d2290314971ef2e2.

* chore(docs): upgrade to mdx3 syntax

* chore(deps): bump @ionic-docs/preset-classic

* fixup(links): fix broken docusaurus v3 links

it appears that in docusaurus v3.0.0 + v3.0.1, links that have a newline
in their public-facing name are considered broken. for the purposes of
the migration, we just remove the newline and unwrap lines that
docusaurus was reporting on `docusaurus build`

* fixup(links): fix broken docusaurus v3 links (v4.9)

it appears that in docusaurus v3.0.0 + v3.0.1, links that have a newline
in their public-facing name are considered broken. for the purposes of
the migration, we just remove the newline and unwrap lines that
docusaurus was reporting on `docusaurus build`

* chore(deps): bump @ionic-docs/preset-classic dep

---------

Co-authored-by: Ryan Waskiewicz <ryanwaskiewicz@gmail.com>