docs(copy-tasks): rewrite documentation for copy tasks
This updates the documentation for the `copy` configuration that can be
made on the `www`, `dist-custom-elements`, and `dist` output targets.
This was motivated by wanting to clarify that this option is not
supported on the documentation-related output targets (e.g.
`docs-json`).

A change was recently made in Stencil to document the related interface,
which we can take advantage of here:
https://github.com/ionic-team/stencil/pull/5413

fixes #852
STENCIl-1047