docs(props): clarify runtime types of union props (#1354)
extend the section on union types to clarify the runtime type of a union
type at runtime. previously, it was difficult to understand/parse the
existing docs to determine how a `@Prop`'s runtime type was determined
when used in a running library or application. this commit intends to
make this a bit clearer, especially in the case of `string|boolean`
types being used in HTML and TSX