docs(support): update typescript support for 5.4.3 (#1381)
update the typescript support table to account for typescript v5.4.3.

the thinking here even though this is a patch version of TS, it reverts
a handful of PRs that made it into a release, whose behavior folks could
be relying on. as a result, I'd rather be safe than sorry here.

update the v4.13.0 table to specify v5.4.2 instead of v5.4.0, as the
typescript team uses the ".0" patch for beta releases and ".1" patch
for RC at the moment.